chore: local, data 패키지 이동
