+++ app/src/main/java/woowacourse/shopping/data/model/Product.kt
@@ -1,4 +1,4 @@
-package woowacourse.shopping.model
+package woowacourse.shopping.data.model
 
 class Product(
     val name: String,
