+++ app/src/main/java/woowacourse/shopping/local/CartProductDao.kt
@@ -1,4 +1,4 @@
-package woowacourse.shopping.data
+package woowacourse.shopping.local
 
 import androidx.room.Dao
 import androidx.room.Insert
