+++ app/src/main/java/woowacourse/shopping/local/CartProductEntity.kt
@@ -1,4 +1,4 @@
-package woowacourse.shopping.data
+package woowacourse.shopping.local
 
 import androidx.room.Entity
 import androidx.room.PrimaryKey
