+++ app/src/main/java/woowacourse/shopping/module/DatabaseModule.kt
@@ -7,8 +7,8 @@ import dagger.Provides
 import dagger.hilt.InstallIn
 import dagger.hilt.android.qualifiers.ApplicationContext
 import dagger.hilt.components.SingletonComponent
-import woowacourse.shopping.data.CartProductDao
-import woowacourse.shopping.data.ShoppingDatabase
+import woowacourse.shopping.local.CartProductDao
+import woowacourse.shopping.local.ShoppingDatabase
 import javax.inject.Singleton
 
 @InstallIn(SingletonComponent::class)
