+++ app/src/main/java/woowacourse/shopping/module/DateFormatterModule.kt
@@ -4,6 +4,7 @@ import dagger.Binds
 import dagger.Module
 import dagger.hilt.InstallIn
 import dagger.hilt.android.components.ActivityComponent
+import dagger.hilt.android.scopes.ActivityScoped
 import woowacourse.shopping.ui.cart.DateFormatter
 import woowacourse.shopping.ui.cart.KoreanLocaleDateFormatter
 
@@ -11,6 +12,7 @@ import woowacourse.shopping.ui.cart.KoreanLocaleDateFormatter
 @InstallIn(ActivityComponent::class)
 abstract class DateFormatterModule {
     @Binds
+    @ActivityScoped
     abstract fun dateFormatter(
         dateFormatter: KoreanLocaleDateFormatter
     ): DateFormatter
