+++ app/src/main/java/woowacourse/shopping/ui/cart/CartActivity.kt
@@ -9,8 +9,10 @@ import woowacourse.shopping.databinding.ActivityCartBinding
 
 @AndroidEntryPoint
 class CartActivity : AppCompatActivity() {
-
     private val binding by lazy { ActivityCartBinding.inflate(layoutInflater) }
+
+    private val cartFragment by lazy { CartFragment() }
+    private val todayFragment by lazy { TodayFragment() }
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
